requirements.txt
decorator==4.0.6
Django==1.9
django-extensions==1.5.9
djangorestframework==3.3.1
FormEncode==1.3.0
funcsigs==0.4
IMDbPY==5.0
lxml==3.5.0
mock==1.3.0
pbr==1.8.1
six==1.10.0
SQLAlchemy==1.0.10
sqlalchemy-migrate==0.10.0
SQLObject==2.1.2
sqlparse==0.1.18
Tempita==0.5.2
wheel==0.24.0
